move interest component  into root routes folder
