+++ frontend/src/routes/Interests.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import Interests from '../../components/Interests';
-import Authlayout from '../../components/AuthLayout';
+import Interests from '../components/Interests';
+import Authlayout from '../components/AuthLayout';
 
 const UserInterest = () => {
   return (
@@ -9,7 +9,8 @@ const UserInterest = () => {
       titleSize={32}
       subTitle={() => (
         <>
-          What are you <span>interested</span> in?
+          What are you <br />
+          <span>interested</span> in?
         </>
       )}
       formAlign="left"
